+++ test/ace/http1/server_test.exs
@@ -296,6 +296,24 @@ defmodule Ace.HTTP1.ServerTest do
     assert request.body == false
   end
 
+  test "handles path beginning with double stash", %{port: port} do
+    http1_request = """
+    GET //foo//bar HTTP/1.1
+    host: example.com:1234
+
+    """
+
+    {:ok, socket} = :ssl.connect({127, 0, 0, 1}, port, [:binary])
+    :ok = :ssl.send(socket, http1_request)
+
+    assert_receive {:"$gen_call", from, {:headers, request, state}}, 1000
+    GenServer.reply(from, {[], state})
+
+    assert request.path == ["foo", "bar"]
+    assert request.raw_path == "//foo//bar"
+  end
+
+  # cant send scheme
   test "handles request with split start-line ", %{port: port} do
     part_1 = "GET /foo/bar?var"