+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## next
+
+### Fixed
+
+- Correctly handle paths beginning with `//`.
+
 ## [0.18.6](https://github.com/CrowdHailer/Ace/tree/0.18.6) - 2019-04-16
 
 ### Added